test-requirements.txt
asgi-lifespan==2.1.0
beanie==1.22.6
colorama==0.4.6
faker==19.6.2
fastapi-users[beanie]==12.1.2
fastapi==0.103.2
flake8==6.1.0
httpx==0.25.0
mypy==1.5.1
pydantic-settings==2.0.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest==7.4.2
tox-gh-actions==3.1.3
tox==4.11.3
uvicorn[standard]==0.23.2
